% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighte_average.R
\name{weighted_average}
\alias{weighted_average}
\title{function to calculate weighted average}
\usage{
weighted_average(value, sample_size)
}
\arguments{
\item{value}{a numeric vector containing the values to be calculated}

\item{sample_size}{a numeric vector containing the n for each value}
}
\value{
one value which added each value together weighted by their n
}
\description{
value should contain the values to be summed up, sample_size
indicates the weight of the value with the same position, the weight is
calculated by dividing the individual sample_size value with the summ of
sample_size
}
\examples{
final_grade <- weighted_average(grades, ETCS)

}
